Update to make the dev branch actually work.
The old pom did not properly compile the jar since it used old links to
maven for example.

It did not update town files from 2.4.0 since 2.4.0 did not contain a
warps section.  Now it does and checks for a town actually existing in
the .meta.yml.  If it doesn't exist then it ignores it and should be
removed on save of it.

Added some help messages to commands.  Need some more work but its a
start.